modify energy interface

+++ src/binary_polynomial_model.hpp
@@ -217,7 +217,7 @@ public:
       poly_value_list_.resize(num_interactions);
       
 #pragma omp parallel for
-      for (int64_t i = 0; i < num_interactions; ++i) {
+      for (int64_t i = 0; i < (int64_t)num_interactions; ++i) {
          std::vector<IndexType> temp;
          for (const auto &it: poly_key_distance_list[i]) {
             temp.push_back(variables[it]);
@@ -586,7 +586,7 @@ public:
       
       if (omp_flag) {
 #pragma omp parallel for reduction (+: val)
-         for (int64_t i = 0; i < num_interactions; ++i) {
+         for (int64_t i = 0; i < (int64_t)num_interactions; ++i) {
             int32_t spin_multiple = 1;
             for (const auto &index: poly_key_list_[i]) {
                spin_multiple *= sample.at(index);
@@ -635,7 +635,7 @@ public:
       
       if (omp_flag) {
 #pragma omp parallel for reduction (+: val)
-         for (int64_t i = 0; i < num_interactions; ++i) {
+         for (int64_t i = 0; i < (int64_t)num_interactions; ++i) {
             int32_t spin_multiple = 1;
             for (const auto &index: poly_key_list_[i]) {
                spin_multiple *= sample_vec[variables_to_integers_.at(index)];
@@ -667,7 +667,7 @@ public:
    PolynomialValueList<FloatType> energies(const std::vector<Sample<IndexType>> &samples) const {
       PolynomialValueList<FloatType> val_list(samples.size());
 #pragma omp parallel for
-      for (int64_t i = 0; i < samples.size(); ++i) {
+      for (int64_t i = 0; i < (int64_t)samples.size(); ++i) {
          val_list[i] = energy(samples[i], false);
       }
       return val_list;
@@ -679,7 +679,7 @@ public:
    PolynomialValueList<FloatType> energies(const std::vector<std::vector<int32_t>> &samples_vec) {
       PolynomialValueList<FloatType> val_list(samples_vec.size());
 #pragma omp parallel for
-      for (int64_t i = 0; i < samples_vec.size(); ++i) {
+      for (int64_t i = 0; i < (int64_t)samples_vec.size(); ++i) {
          val_list[i] = energy(samples_vec[i], false);
       }
       return val_list;
@@ -868,7 +868,7 @@ public:
       std::vector<IndexType> sorted_variables = get_sorted_variables();
       
 #pragma omp parallel for
-      for (int64_t i = 0; i < num_interactions; ++i) {
+      for (int64_t i = 0; i < (int64_t)num_interactions; ++i) {
          std::vector<std::size_t> temp;
          for (const auto &it: poly_key_list_[i]) {
             auto it_index = std::lower_bound(sorted_variables.begin(), sorted_variables.end(), it);
